requirements.txt
# This file lists the essential runtime dependencies for the main backend application,
# which includes the CRUD API, the AI chat service, and the MCP server.
# Development-only packages like 'pytest' have been excluded.

# --- Core Web Framework ---
# FastAPI is the main web framework.
# Uvicorn is the ASGI server to run the application. [standard] includes performance extras.
fastapi
uvicorn[standard]

# --- Database Layer ---
# SQLAlchemy is the ORM used by the main CRUD app.
# Alembic is used for database migrations.
# Psycopg2 is the synchronous driver for PostgreSQL, required by Alembic.
# AsyncPG is the asynchronous driver for PostgreSQL, used by FastAPI.
# Pydantic-Settings is used for managing configuration from environment variables.
sqlalchemy
alembic
psycopg2-binary
asyncpg
pydantic-settings

# --- AI & Real-Time Communication (WebRTC) ---
# The core library for handling WebRTC connections.
# This library has many dependencies (cryptography, av, etc.) which pip will install.
aiortc
livekit
livekit-api   
# AV is a direct and critical dependency of aiortc for handling media frames.
# It's good practice to list it explicitly.
av

# --- External Communication ---
# A modern, async-capable HTTP client used by the AI to call the CRUD API
# and the external LM Studio API. Also used by the MCP server.
httpx

# --- Model Context Protocol (MCP) ---
# The library for running the MCP server to expose tools to LM Studio.
mcp
pipecat-ai[azure,livekit,silero,cartesia]